+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - resource `stream_library`: update default player version;
 - resource `compute_container_app`: standardize volume name validation;
 - resource `pullzone`: improve `origin` block validation;
+- resource `pullzone_edgerule`: validate action type;
 
 ## 0.13.2 - 2026-03-10